move router mutation to actions
