+++ src/components/Form/ChipList.jsx
@@ -2,31 +2,17 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
 import Chip from 'material-ui/Chip';
-import querystring from 'querystring';
 // import { form } from 'reducers';
-// import { deleteChip } from 'actions';
+import { deleteChip } from 'actions';
 import styles from './ChipList.css';
 
-const ChipList = ({ chipList, router, name }) => (
+const ChipList = ({ deleteChip, chipList, router, name }) => (
   <div className={styles.container}>
     {chipList.map((chip, index) => (
       <Chip
         style={{ textTransform: 'uppercase', margin: '0 5px 5px 0' }}
         key={index}
-        onRequestDelete={() => {
-          const query = querystring.parse(window.location.search.substring(1));
-          const field = [].concat(query[name] || []);
-          router.push({
-            pathname: window.location.pathname,
-            query: {
-              ...query,
-              [name]: [
-                ...field.slice(0, index),
-                ...field.slice(index + 1),
-              ],
-            },
-          });
-        }}
+        onRequestDelete={() => deleteChip(router, name, index)}
       >
         {chip.text}
       </Chip>
@@ -36,6 +22,8 @@ const ChipList = ({ chipList, router, name }) => (
 
 const mapStateToProps = () => ({});
 
-const mapDispatchToProps = () => ({});
+const mapDispatchToProps = (dispatch) => ({
+  deleteChip: (router, name, index) => dispatch(deleteChip(router, name, index))
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(withRouter(ChipList));
